add malloc protection and return with put error, exept in execution file and search_cmd

+++ parsing/utils_cmd_identify.c
@@ -6,7 +6,7 @@
 /*   By: opdi-bia <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 16:18:49 by opdi-bia          #+#    #+#             */
-/*   Updated: 2024/10/08 16:22:39 by opdi-bia         ###   ########.fr       */
+/*   Updated: 2024/10/09 17:16:22 by opdi-bia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,18 +31,22 @@ char    *check_line(t_data *data, char *buffer, char *delimiter, int cmd)
     write(data->token[cmd].fdin, buffer, ft_strlen(buffer) + 1);
     write(data->token[cmd].fdin, "\n", 2); 
     buffer = readline(">");
+    if(buffer == NULL)
+        return(put_error(ERR_MALLOC, NULL), NULL);
     return(buffer);
 }
 
-void    interrupt_heredoc(t_data *data, int new, int cmd)
+int    interrupt_heredoc(t_data *data, int new, int cmd)
 {
-    dup2(new, STDIN_FILENO);
+    if(dup2(new, STDIN_FILENO) == -1)
+        return (perror("dup2"), -1);;
     close(data->token[cmd].fdin);
     unlink("temp_file_gere_doc.txt");
     g_sig_recieved = 0;
+    return(0);
 }
 
-void    check_arg(t_data *data, int i, e_type type)
+int    check_arg(t_data *data, int i, e_type type)
 {
     int j;
     int cmd;
@@ -60,11 +64,18 @@ void    check_arg(t_data *data, int i, e_type type)
             if(data->token[cmd].nb_arg + 1 > data->token[cmd].size)
             {
                 data->token[cmd].litteral = my_realloc(data->token[cmd], data->token[cmd].size);
+                if(data->token[cmd].litteral == NULL)
+                    return(put_error(ERR_MALLOC, NULL), -1);
                 data->token[cmd].size += 1;
             }
             data->token[cmd].litteral[j] = ft_strdup(data->token[i].litteral[0]);
+            free(data->token[i].litteral[0]);
+            if(data->token[cmd].litteral[j] == NULL)
+                return(put_error(ERR_MALLOC, NULL), -1);
             j++;
         }
         i++;
     }
+    return(0);
 }
+
